[components] Fix popover scrolling (#343)
