+++ packages/@sanity/components/sanity.json
@@ -541,11 +541,11 @@
     },
     {
       "implements": "part:@sanity/components/edititem/popover",
-      "path": "edititem/PopOver.js"
+      "path": "edititem/EditItemPopOver.js"
     },
     {
       "implements": "part:@sanity/components/edititem/fold",
-      "path": "edititem/FoldOut.js"
+      "path": "edititem/EditItemFoldOut.js"
     },
     {
       "implements": "part:@sanity/components/previews/default",
@@ -741,11 +741,11 @@
     },
     {
       "implements": "part:@sanity/components/edititem/popover-style",
-      "path": "edititem/styles/PopOver.css"
+      "path": "edititem/styles/EditItemPopOver.css"
     },
     {
       "implements": "part:@sanity/components/edititem/fold-style",
-      "path": "edititem/styles/FoldOut.css"
+      "path": "edititem/styles/EditItemFoldOut.css"
     },
     {
       "implements": "part:@sanity/components/previews/default-style",
