+++ packages/@sanity/components/src/edititem/styles/EditItemFoldOut.css
